Add a grantsd binary.
Add a basic grantsd binary that just exposes the version information.

This needs the darlinggo.co/version package, which is now vendored.

We also created a Dockerfile, and vendored a ca-certificates.crt file
for it, so it can resolve HTTPS requests.

Fix our build script, which had an embarrassing copy/paste error in it.