Add a health check.
Address #4 by creating a health check endpoint for the grantsd
binary. This is mostly a matter of importing the Darling
healthcheck library.

The health check verifies that we have a connection to Postgres
and that we can run a query against it. That’s all it does, for
the moment.