Start work on the API.
Flesh out the basis of the API. It largely revolves, right now,
around a single handler. That handler accepts requests for tokens,
and uses a granter interface to check to see if the request has a
valid grant.

This commit comes with a refreshTokenGranter, which implements the
logic for the refresh_token grant type. Or, at least, it should. I
haven’t had a chance to really extensively test it just yet.

The idea is that we call out to the granters for the stuff that is
unique to them:

* Validating the request is a valid grant, whatever that means to
  the granter. For refresh tokens, it means that the token is
  valid.
* Creating a Grant object with the necessary audit information.
* Taking any action after the grant type has been successfully
  used; in the case of refresh tokens, marking it as used so it
  can’t be used again.
* Letting us know whether or not we should be redirecting to
  another URL (implicit and authorization code grants) to deliver
  the token.

The rest is the responsibility of the HTTP handler, which
coordinates everything. This allows us to implement grant types
without redoing the same work a billion times.