Drop vendor, validate tokens.
Get rid of vendor, since this now an imported library and not something
that is run as its own binary.

Validate our refresh tokens and access tokens using the new monolith
model. This largely just means that the grants package now depends on
the tokens and (new) sessions package, and calls through to their
Storers or helper functions as necessary.