Comment our API functions, actually return stuff.
Add a bunch of comments to the API functions so it's easier to follow
what's going on with all those helper libraries.

Implement token and error writing and redirecting (fixes #8 and
fixes #9). This brought in a new TODO to figure out what the client's
URI should be in the event they don't supply one or supply a partial.
We're going to have to implement that client package soon; it's getting
harder and harder to code around.

Because our query params are handled in an additive way, state is
automatically preserved if it's supplied (fixes #17).