Add Google Cloud Datastore as a Storer.
Update our Grant type to not use pqarrays, and instead translate it into
a PostgreSQL-specific type before inserting it into the database and
after retrieving it from the database.

Add a Google Cloud Datastore implementation of the Storer, which should
not have any eventual consistency concerns.

Fix a typo in the apiv1 package that was preventing it from building.