Upgrade Grants to be able to replace refrehs tokens.
If we add the concept of ancestry to Grants (so we can see when a Grant
is generated by another Grant and trace that chain back to the
beginning) and make Grants revokable, they can replace the entirety of
our tokens package. Refresh tokens will just be grants that are issued
to the client and exchanged for a new access token and new grant, then.

This allows us to avoid the awkward situation where a refresh token gets
exchanged for a grant and we need to mark both the grant and refresh
token as used. Without tightly coupling them, that setup has to allow
for an edge case where one gets marked as used but the other doesn't, if
the second to get marked as used returns an error. While it's possible
to sort of work around this (mark grant as used, then mark the refresh
token as used, because the refresh token can then be re-exchanged for a
new grant) it's weird and could lead to some confusion. By combining
them, we avoid the headache altogether and get to delete some code,
which is always nice.