Update Postgres to use a separate table for Ancestors.
Storing all our ancestor IDs in a string array is tempting but probably
a bad idea. In theory, there's no upper bound to the number of items in
the array, and at 36 bytes apiece, it takes about a year of normal
refreshing (once every two weeks) to get to storing 1KB per grant, which
probably isn't the end of the world but certainly doesn't seem optimal.
And while the storage requirements don't change (actually, increase)
when storing them in a separate table, querying against them is much
simpler and introducing paging later is also more straightforward.